+++ src/main/java/com/sprint/mission/discodeit/dto/auth/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.auth.request;
+package com.sprint.mission.discodeit.dto.auth;
 
 public record LoginRequest(
         String username,
