+++ src/main/java/com/sprint/mission/discodeit/dto/message/MessageUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.message.request;
+package com.sprint.mission.discodeit.dto.message;
 
 public record MessageUpdateRequest(
         String newContent
