+++ src/main/java/com/sprint/mission/discodeit/dto/readstatus/ReadStatusCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.readstatus.request;
+package com.sprint.mission.discodeit.dto.readstatus;
 
 import java.time.Instant;
 import java.util.UUID;
