+++ src/main/java/com/sprint/mission/discodeit/dto/user/UserDto.java
@@ -1,0 +1,13 @@
+package com.sprint.mission.discodeit.dto.user;
+
+import com.sprint.mission.discodeit.dto.binarycontent.BinaryContentDto;
+
+import java.util.UUID;
+
+public record UserDto(
+        UUID id,
+        String username,
+        String email,
+        BinaryContentDto profile,
+        Boolean online
+) {}
